Move CapturingViewFactory into a separate file
