+++ Tests/SteamPressTests/BlogControllerTests.swift
@@ -318,103 +318,3 @@ class BlogControllerTests: XCTestCase {
         XCTAssertEqual(nodeArray[1]["name"]?.string, "The second tag")
     }
 }
-
-import URI
-import Foundation
-
-class CapturingViewFactory: ViewFactory {
-    
-    private func createDummyView() -> View {
-        return View(data: "Test".makeBytes())
-    }
-
-    func createBlogPostView(uri: URI, errors: [String]?, title: String?, contents: String?, slugUrl: String?, tags: [Node]?, isEditing: Bool, postToEdit: BlogPost?, draft: Bool) throws -> View {
-        return createDummyView()
-    }
-
-    func createUserView(editing: Bool, errors: [String]?, name: String?, username: String?, passwordError: Bool?, confirmPasswordError: Bool?, resetPasswordRequired: Bool?, userId: Identifier?, profilePicture: String?, twitterHandle: String?, biography: String?, tagline: String?) throws -> View {
-        return createDummyView()
-    }
-
-    func createLoginView(loginWarning: Bool, errors: [String]?, username: String?, password: String?) throws -> View {
-        return createDummyView()
-    }
-
-    private(set) var adminViewErrors: [String]? = nil
-    func createBlogAdminView(errors: [String]?) throws -> View {
-        adminViewErrors = errors
-        return createDummyView()
-    }
-
-    private(set) var resetPasswordErrors: [String]? = nil
-    private(set) var resetPasswordErrorFlag: Bool? = nil
-    private(set) var resetPasswordConfirmErrorFlag: Bool? = nil
-    func createResetPasswordView(errors: [String]?, passwordError: Bool?, confirmPasswordError: Bool?) throws -> View {
-        self.resetPasswordErrors = errors
-        self.resetPasswordErrorFlag = passwordError
-        self.resetPasswordConfirmErrorFlag = confirmPasswordError
-        return createDummyView()
-    }
-
-    private(set) var author: BlogUser? = nil
-    private(set) var isMyProfile: Bool? = nil
-    private(set) var authorPosts: Page<BlogPost>? = nil
-    private(set) var authorURI: URI? = nil
-    func createProfileView(uri: URI, author: BlogUser, isMyProfile: Bool, paginatedPosts: Page<BlogPost>, loggedInUser: BlogUser?) throws -> View {
-        self.author = author
-        self.isMyProfile = isMyProfile
-        self.authorPosts = paginatedPosts
-        self.authorURI = uri
-        return createDummyView()
-    }
-
-    private(set) var blogPost: BlogPost? = nil
-    private(set) var blogPostAuthor: BlogUser? = nil
-    private(set) var blogPostURI: URI? = nil
-    func blogPostView(uri: URI, post: BlogPost, author: BlogUser, user: BlogUser?) throws -> View {
-        self.blogPost = post
-        self.blogPostAuthor = author
-        self.blogPostURI = uri
-        return createDummyView()
-    }
-
-    private(set) var tag: BlogTag? = nil
-    private(set) var tagPosts: Page<BlogPost>? = nil
-    private(set) var tagUser: BlogUser? = nil
-    private(set) var tagURI: URI? = nil
-    func tagView(uri: URI, tag: BlogTag, paginatedPosts: Page<BlogPost>, user: BlogUser?) throws -> View {
-        self.tag = tag
-        self.tagPosts = paginatedPosts
-        self.tagUser = user
-        self.tagURI = uri
-        return createDummyView()
-    }
-
-    private(set) var blogIndexTags: [BlogTag]? = nil
-    private(set) var blogIndexAuthors: [BlogUser]? = nil
-    private(set) var paginatedPosts: Page<BlogPost>? = nil
-    private(set) var blogIndexURI: URI? = nil
-    func blogIndexView(uri: URI, paginatedPosts: Page<BlogPost>, tags: [BlogTag], authors: [BlogUser], loggedInUser: BlogUser?) throws -> View {
-        self.blogIndexTags = tags
-        self.paginatedPosts = paginatedPosts
-        self.blogIndexURI = uri
-        self.blogIndexAuthors = authors
-        return createDummyView()
-    }
-    
-    private(set) var allAuthorsURI: URI? = nil
-    private(set) var allAuthorsPageAuthors: [BlogUser]? = nil
-    func allAuthorsView(uri: URI, allAuthors: [BlogUser], user: BlogUser?) throws -> View {
-        self.allAuthorsURI = uri
-        self.allAuthorsPageAuthors = allAuthors
-        return createDummyView()
-    }
-    
-    private(set) var allTagsURI: URI? = nil
-    private(set) var allTagsPageTags: [BlogTag]? = nil
-    func allTagsView(uri: URI, allTags: [BlogTag], user: BlogUser?) throws -> View {
-        self.allTagsURI = uri
-        self.allTagsPageTags = allTags
-        return createDummyView()
-    }
-}
